ext4: set default journal ioprio to medium real time priority
kjournald2 is the kernel thread on which most other processes depend on
for doing the filesystem portion of their IO. As such, there exists
the equivalent of a priority inversion situation, where kjournald2
would get less priority as it should.

For this reason we set to "medium real time priority", which is higher
than regular tasks, but not infinitely powerful.

Signed-off-by: xiongping1 <xiongping1@xiaomi.com>
Change-Id: I58d026900175b5bf8b75050b13813db9643c89cb